feat: introduce group manager
Moved group methods to group package
Moved createTestControllerEnvironment function to jimmtest package
