+++ internal/jimm/group/export_test.go
@@ -1,0 +1,6 @@
+// Copyright 2024 Canonical.
+
+package group
+
+// GroupManager is a type alias to export groupManager for use in tests.
+type GroupManager = groupManager
