+++ internal/jujuapi/controllerroot.go
@@ -28,7 +28,6 @@ import (
 )
 
 type JIMM interface {
-	GroupService
 	RelationService
 	ControllerService
 	LoginService
@@ -52,6 +51,7 @@ type JIMM interface {
 	GetCloudCredentialAttributes(ctx context.Context, u *openfga.User, cred *dbmodel.CloudCredential, hidden bool) (attrs map[string]string, redacted []string, err error)
 	GetCredentialStore() credentials.CredentialStore
 	GetRoleManager() jimm.RoleManager
+	GetGroupManager() jimm.GroupManager
 	GetJimmControllerAccess(ctx context.Context, user *openfga.User, tag names.UserTag) (string, error)
 	// FetchIdentity finds the user in jimm or returns a not-found error
 	FetchIdentity(ctx context.Context, username string) (*openfga.User, error)
