+++ internal/testutils/jimmtest/jimm_mock.go
@@ -30,7 +30,6 @@ import (
 // a NotImplemented error.
 type JIMM struct {
 	mocks.RelationService
-	mocks.GroupService
 	mocks.ControllerService
 	mocks.LoginService
 	mocks.ModelManager
@@ -53,6 +52,7 @@ type JIMM struct {
 	GetCloudCredential_                func(ctx context.Context, user *openfga.User, tag names.CloudCredentialTag) (*dbmodel.CloudCredential, error)
 	GetCloudCredentialAttributes_      func(ctx context.Context, u *openfga.User, cred *dbmodel.CloudCredential, hidden bool) (attrs map[string]string, redacted []string, err error)
 	GetCredentialStore_                func() jimmcreds.CredentialStore
+	GetGroupManager_                   func() jimm.GroupManager
 	GetRoleManager_                    func() jimm.RoleManager
 	GetJimmControllerAccess_           func(ctx context.Context, user *openfga.User, tag names.UserTag) (string, error)
 	FetchIdentity_                     func(ctx context.Context, username string) (*openfga.User, error)
@@ -216,6 +216,13 @@ func (j *JIMM) GetRoleManager() jimm.RoleManager {
 		return nil
 	}
 	return j.GetRoleManager_()
+}
+
+func (j *JIMM) GetGroupManager() jimm.GroupManager {
+	if j.GetGroupManager_ == nil {
+		return nil
+	}
+	return j.GetGroupManager_()
 }
 
 func (j *JIMM) GetJimmControllerAccess(ctx context.Context, user *openfga.User, tag names.UserTag) (string, error) {
